agda: don't install Everything module
The Everthing module is not part of a library and should therefore
not be copied to the nix store.

This is particularly bad, if the Everything module is defined in
an agda library included directory, e.g. consider a agda-lib with
`include: .` with Everything.agda in the project root. In this case
the Everything file would become part of the library.

If multiple libraries inlcude the project root the Everthing module
would now be ambigous and the build would fail.